Move the three x,y,z points columns in the csv file to the begining
The file written by the VTK CSV writer has the x,y,z points columns at
the end.  This new code will move the columns to the begining at the cost
if an extra file read & write.  I looked at the implementation of the
CSV writer in ParaView and it did not seem easy to extend it to allow
repositioning of the points columns.