fix: Temporary fix for dump ranges
This is an additional fix for the patch that removed support for
converting fpos_t to integer.

First process the skip information to make sure that we dont get

For some reason packets will occasionally have the first six HDL
firings zeroed out.  Skipping over these gives a more reasonable
method of counting the number of frames.

I am not sure why ReadFrameInformation does not need a similar test.

Remaining issue: We will almost always get a sliver of a frame as the
first frame using this method, because we write the whole packet that
the frame starts in.