enh: Add interpolation of GPS position to high-resolution timer
This builds an interpolation function on the GPS location and time.
The UTC time of the GPS values are converted to microseconds from top
of the hour to match timing signal of IMU data.

Assumptions
1) The minutes and seconds UTC of the GPS signal can be corresponded to
the microsecond since top of the hour of the packet timestamp.

2) There is no accounting for the potential wraparound of datasets of
microseconds since top of the hour.  Data which contains a values near
the end of the hour and near the beginning will not behave correctly.