improve "unit test"
Teach the test executable a second mode where it takes a 'test script'
that contains a list of 'test commands' to execute. These are things
like 'load frame <N>', 'assert that point count is <X>', etc.. This will
allow us to provide something like an actual unit test, without using a
hard-coded test data file or hard-coded assertions. (Actual data files
and/or test scripts will come later.)

Note that there is a not-yet-implemented 'vertex' command that will
provide a way to test the expected value(s) of a particular point.