add 'dual return' flags to reader
Add an additional data field with 'dual return' flags, which will allow
visualizing which returns are unique and which are part of a two-return
pair, and of the pairs, separating the pair by either intensity or
distance. Add code to detect when a dual return occurs and to set these
flags accordingly.