add 'dual return' deltas to reader
Also make available the mean-difference values for distance and
intensity as data fields for dual return values. This provides a crude
(but usable) mechanism for visualizing only one of the attributes (and
without a convoluted color map), by picking the interesting difference
and setting the mapping range very low. (Alternatively, it provides a
way to visualize the continuous range of difference between returns.)