overhaul wrapping of vvSelectFramesDialog
Remove all the manual wrapping work-arounds for vvSelectFramesDialog and
wrap it the way PythonQt intends (and is better API anyway); using slots
and Q_PROPERTY. This also exposes the enums to Python, which lets us use
symbolic names for the frame mode rather than "magic numbers". Also
refactor getFrameSelectionFromUser to return an object with properties
rather than a long tuple.