first pass implementation of LAS writing
Create a new class to handle writing to LAS files. Finish connecting the
UI to this new writer. This "works" in the sense that it produces a file
without crashing, but there are very likely issues with the exact manner
in which the data is being written (scaling issues, probable need to
offset the position information according to the GPS information, etc.).