map split flags to "logical" values
Change how we populate the split dual-return flag fields to have
"logical" values {-1, 0, 1} (for low/near, single-return, and high/far,
respectively) rather than being just a straight bitmask of the raw bit
values. This way, the values are more meaningful if a user is actually
looking at the numeric values, and also puts single-return results in
the middle of the mapping range. (In a way, this is a "binned" version
of the difference fields that we used to have.)