add sensor transform options
Extend the calibration dialog to also include the ability to specify the
sensor position and orientation. Do some general clean-up of the dialog
while we're at it. Move invocation of the dialog to applogic.py so that
we can retrieve the transform (as it can't readily be passed as a string
to runPython()). Apply transform when reading points.