implement Applanix import
Add a new 'open' mode that takes an Applanix position data file in
addition to the usual Wireshark capture of the Velodyne sensor (LiDAR)
output. Seriously refactor / clean up pqVelodyneManager while we're at
it and needing to fiddle with it anyway.

This works by feeding openPCAP two file names, one with the LiDAR data
and one with the position data. Note that this doesn't work with recent
files, currently (only the pcap file will be remembered).