apply Applanix geoposition transform
Set up geoposition transformations and an interpolator for the same for
vtkApplanixPositionReader. Modify vtkVelodyneHDLReader and applogic.py
to use these. (Note that this breaks use of vtkVelodyneHDLPositionReader
until that is adapted to do likewise. However, it also fixes applying
the geoposition heading prior to the sensor transform. Note also that
there is a lot of hackish and hard-coded time manipulation logic at the
moment.)