partly implement transformed output
Add a transform option when exporting to CSV/LAS. Right now, this allows
selection at export time between always using sensor-relative
coordinates or using relative geospatial coordinates (i.e. taking the
sensor geospatial position into account). This could previously be done
anyway by having the appropriate display mode selected, but that could
easily confuse users, as it is not obvious that that setting would
affect the output. (Essentially what we are doing currently is to
override it during the export.)

Note that the 'absolute geoposition' modes are not yet implemented for
any export, and probably will not work for CSV for some time.