implement LAS absolute UTM mode
Modify LAS writer to accept a UTM origin specification, which causes it
to (attempt to) write an SRS and offset, and apply the offset to the
relative positions such that the resulting LAS file provides absolute,
properly georeferenced positions. Also add a time range option, which is
used to limit output to frames for which a geotransform is available.
Modify the various support logics to use these features appropriately.

(Note that writing the SRS currently fails, as it requires liblas to be
built with GDAL, which we are not currently doing.)