implement GCS conversion in LAS writer
Add a mechanism to vtkLASFileWriter to allow specifying the desired
output GCS (other than UTM). Use this to implement lat/lon output.

Note that this is only partly implemented for PROJ 4.7, due to the
appallingly poor API of the same; the only supported output GCS is 4326
(WGS'84), and the input must be a WGS'84-related UTM zone. There is also
code for PROJ 4.8 which supports any input and output GCSs.