add parameter for time offset correction
Add a TimeOffset property to vtkApplanixPositionReader to allow
configuring the time offset that is applied to correct for the offset
between GPS time and UTC time. This makes it "technically" configurable,
but there is no UI yet, nor is it being used in the application logic
(and also the ParaView wrapper likely will need to be changed as well).