refactor time handling
Remove the various 'top of hour' conversions. Refactor time handling in
vtkVelodyneHDLReader to use the position interpolator time range (which,
for Applanix, gives time-of-week) to correct the point data timestamps
in order to match the interpolator and compensate for wrapping at the
top of the hour (from the point timestamps giving milliseconds since top
of hour).