update vtkVelodyneHDLPositionReader
Refactor vtkVelodyneHDLPositionReader to work with the new mechanisms
for applying geoposition transformation to the sensor points, which
expects to get vtkTransform's out of the interpolator rather than
tuples. Note that the computing reinterpolated position and heading
values has been dropped; this isn't necessary for transforming the LiDAR
points, and will only affect exported position information (which will
more closely match the input information as a result).