fix Eigen build with ninja
Remove the broken ("insufficient" might be more accurate) rule to
"build" Eigen by just copying over the header directory, and instead
patch the build with the upstream version of language_support.cmake that
fixes the fortran detection bug that was preventing us from just
building it normally with Ninja. (Note that "upstream" here means "not
yet in a released version", or we wouldn't be having this problem in the
first place.)

This is needed for any users that require the 'unsupported' headers for
Eigen, which were not being "installed". Note that the include directory
was also "wrong" (didn't match what the real build does); anyone using
$<INSTALL>/include/eigen must use $<INSTALL>/include/eigen3 instead.