Add timestamp offsets to individual returns for HDL-32 and VLP-16
This adds individual timestamp offsets for the HDL-32 and VLP-16
sensors.

The azimuth values for individual returns are corrected by
interpolating between azimuths based on the time stamp offsets.

Specificially this works as follows.  The azimuth change
(azimuth_diff) is taken to be the azimuth change between the last
firing and the current (crossing packet boundaries).  (Note the first
firing does not have a last azimuth and so is not interpolated).

For each laser the time of the specific return is divided by the total
time for the firing (producing a value 0..1).  This is used to
interpolate between the given azimuth and azimuth + azimuth_diff.

This affects the X,Y location of points in the tests.