Ensure data is only read once on load
This reorders operations in data loading to ensure that setting of laser
selection and crop parameters do not trigger multiple reads of the data.
This should now be organized to set all the parameters before the first call
to Show.

As a side-effect the behavior to abort load if data did not have an "intensity"
array was removed.  I'm not sure this added a lot of value.