Reload the UI controls to default when clearing settings.
This modifies the way VeloView saves and restores its MainWindow settings.
On reset, VeloView restores a state named "OriginalMainWindow" before
clearing its settings, as clearing the settings alone won't update the UI
and the current MainWindow state will be saved when restarting VeloView.
VeloView restarts right after.

The "OriginalMainWindow" state is a state saved into the settings when:
* Launching VeloView for the first time
* Launching VeloView with a "bad" settings file (i.e. a file without an
"OriginalMainWindow" state)
* Right after resetting the settings

In the second case, a warning message is displayed to the user to prevent
them that the settings were wrong and will be cleared.

As resetting VeloView settings delete old settings and restart the
application, a message box is displayed to warn the user and force them
to confirm that they want to reset the settings.

This fixes issue #14