Add chart view for slam output and "dokerize" them
We now have a lot of different views (such as the SpreadSheetView, the
OverHeadView, and right now the ChartViews). We can't continue to stack
them in this short space at the right of the screen.

That's why we "dockerize" them, now the user can exploit multiscreens by
moving around these different views, and resize them.