fix: dropped feature "GPS listening" in refactoring
This feature was present in 380f39, dropped in fc956d.
(Listening and recording of Velodyne network packets
that relay GPS GRMC sentences from a GPS connected to Lidar).

The problem should be solved according to my manual test.
I tested using the PacketFileSender binary to stream the
VLP32 record file "HDL32-V2_R into Butterfield into Digital Drive.pcap"
This record contains GPS data in the 512 bytes packets.
When the stream is recorded to a new pcap,
and the new pcap is then opened the GPS trace appears in window
"Show overhead view of GPS position".

Currently there is no live-filling of a GPS interpolator (or
vtkPolydata), but to my knowledge this was not possible in the past
either.