Fix race condition crash on lidar packet reception
This crash is due to the fact that vtkSmartPointer is not thread-safe,
so it is possible that two threads enter the destructor of the same
vtkSmartPointer at the same time and both get to a point where the ref
counter is zero, so both try to free it and the second tries to
dereference nullptr / produces a double free.

There is no need to start a recording to reproduce the crash, only to
listen to a Lidar.
Without gdb attached, in RelWithDebInfo builds, the crash occurs
randomly (a few hours are often required).
With gdb attached, in RelWithDebInfo builds, the crash occured faster
(less than 10 minutes).
The crash could not be reproduced in Debug builds.
"Fuzzing" (replaying PCAP at high speeds with PacketFileSender
with stops and restarts at random instants) was useful to trigger the
bug.

A cleaner fix would require modifying many parts of the network code,
which does not seem to be good idea because it might be rewritten in
the near future.

Call trace (first type, more frequent):
0  0x0000000000000000 in ?? ()
1  vtkGarbageCollectorToObjectBaseFriendship::UnRegisterBase
2  vtkGarbageCollectorImpl::FlushEntryReferences
3  vtkGarbageCollectorImpl::CollectInternal
4  vtkGarbageCollector::Collect
5  vtkSmartPointer<vtkPolyData>::~vtkSmartPointer
6  vtkLidarStream::RequestData

Call trace (second type):
0  0x0000000000000000 in ?? ()
1 vtkGarbageCollectorToObjectBaseFriendship::UnRegisterBase
2  vtkGarbageCollectorImpl::FlushEntryReferences
3  vtkGarbageCollectorImpl::CollectInternal
4  vtkGarbageCollector::Collect
5  vtkSmartPointer<vtkPolyData>::~vtkSmartPointer
6  __gnu_cxx::new_allocator<vtkSmartPointer<vtkPolyData>>::destroy
7  std::allocator_traits<std::allocator<
     vtkSmartPointer<vtkPolyData>>>::destroy
8  std::deque<vtkSmartPointer<vtkPolyData>,
     std::allocator<vtkSmartPointer<vtkPolyData>>>::pop_front
9  PacketConsumer::UpdateDequeSize