Revert "Add ninja for simpler build"
This reverts commit e762686d3978102c063fbc668704e1f1fd6d2051.

Reasons not to copy ninja.exe to build directory:
1) it does not make the build much simpler: you still have to install
git, cmake and Visual Studio. So downloading ninja.exe and adding it to
your path is not a lot of work (If you are lazy place it in
C:\Windows\System32). That takes approximately 1 minute.
2) it does not add ninja.exe to the path, so if you want to rebuild
only veloview (for a faster build) you go to
common-superbuild/veloview/build (as indicated by the developer guide),
and in this directory there is no ninja.exe so your command with
"ninja" fails and you are perplexed.
3) it complexifies the work of the CI scripts