fix boken PCAP loading progress dialog
That was broken since fc956de3d "tream mechanism refactoring".
Two problems were appearing:
- the progress dialog was not refreshed (thus it was black) because
the applogic.py callback onProgressEvent() was not triggered anymore
by UpdateProgress()
- when closing, the progress dialog would not give the focus back
to the main window, thus breaking the shortcut unless the user
interacted with the main window at the windows management level,
which was inconvenient.

It was also possible to solve only the second problem by not marking
the progress dialog as 'modal'.