bugfix in stream mode some race condition occured
At each new frame to display, the Interpreter was reloading the CalibrationFile.
This was the cause for incomplete frame at high frequency (>20Hz)
This is a quick bugfix, the Interpreter is still thread unsafe. It is used
in the UI thread to set some parameter and in the Consumer
thread to process packets. So if some parameters are modified while processing
a packet the frame content could be invalid.