bugfix in TrailingFrame
The bug appear when you select the trailingFrame in the spreadsheet view
that is in the QDockWidget. Then wen when changing the frame, for some unknown
reason, multiple call to ResquestDataExtent are executed before the call to
RequestData. This caused the filter to go inside an infinity loop.

The bug seems like a paraview bug, as creating a spreadsheetview with the
pqObjectBuilder directly or doing it inside the UI via the pqMultiTabbedViewWidget
does not result in the same behavior.