Do not use VTK wrapping in plugins
This is required for now as wrapping is not working as-is,
because the macro WRAP_PLUGIN_FOR_PYTHON needs access to the mother class
declaration, which may not be part of the plugin.

This breaks some features and should be fixed when refactoring is done,
possibly by removing the need for this wrapping in the plugins.