Speedup by disabling some arrays if no spreadsheet
This will speedup both pcap reading and stream listening, when
spreadsheet is hidden.

Some arrays are now classified as "Advanced". The property panel gives
access to this option.
Indeed, it makes little sense to propagate some arrays for which it is
hard to give a colorscheme (for instance binary flags).

This "Advanced" option is automatically enabled:
- when the spreadsheet is opened
- when a vtkLidarProvided is instantiated while the spreadsheet is open