Select Lidar Data by default
The Position Orientation Stream was instantiate after the Lidar Stream.
So in the pipeline it was shown and selected by default.
This was confusing for the user because the default coloring matched with the PosOr coloring.
A lambda user doesn't used the pipeline browser so we need to restore the default behavior.

The Position Orientation data are hidden by default (the user have to use the pipeline browser to see them)