[fix] Ensure python paths to be consistent.
We encountered an issue where some cmake python paths was not the same
between lidarview, paraview and pythonqt projects. This was due to some
python variables modified when find_package(PythonsLibs) was called.
That's why we just include FindPythonLibs to get the function
PYTHON_ADD_MODULE needed in WRAP_PLUGIN_FOR_PYTHON.