[ui] Extend the width of the pcap file name area
'GetSensorInformation' was called to display the sensor information part to the user (in the status bar).
We need to call the new function 'GetSensorInformationShort' to reduce this part.
We can now allow more space to the pcap file name.
The maximum size of the pcap file name is changed from 43 characters to 120 characters.