[ui] Change the way the Select Points Through is defined
We change the way we add the Select Point Through action.

Before this commit :
A new action was created in vvMainWindow.ui and called the paraview reaction
  used when you click on the paraview button "Select Points Through" in the view.

The shortcut associated to the paraview action was 'g'.
But if you click 'g' you have access to the paraview functionnality, the paraview button is enable.
But the lidarView button is still disable.
This is confusing for the user.
The shortcut 'g' have to enable the lidarview button.
We can not rewrite a shortcut.
So a simple way is to "ride up" the real paraview button.