[feat] Add a behavior to color the data by the right array according to the sensor for LidarStream
We want to colorize by "intensity" if the Sensor send packets in legacy mode
and colorize by "reflectivity" if the sensor send packets in Advanced mode.

The only way to know the type of sensor is to received packets.
This behavior allows to colorize the data with the right array when the stream receives data.