[fix] Strange apply behavior
Before, after the creating a source, this source was set as Modified by default.
So next time you click on `Apply` on the property panel, it will try to apply
the parameters to the Reader/Stream.
Here it was creating a new view to display the Calibration of the
LidarReader/Stream the first time you click on Apply.

For information, in the pipeline browser, modified source are in bold.