[fix] Move "Update pipeline" call and the removal of the progressbar
This commit is needed to support the switch between APF and Legacy format
The switch is detected in the Meta interpreter in the IsLidarPacket function.
So the switch is detected when we call a "Request Information"
As soon as a switch is detected a calibration dialog should pop up to let the user select the new calibration file.
This new calibration file freezes if the previous calibration dialog is still used.
So we need to move the call to "UpdatePipeline" (which do a Request information) to call it after the last use of the first dialog.
In this case the update pipeline is called when we create the trailing frame on the output of the lidarReader