[feat] Add Behavior to handle the switch of the interpreter
This commit is needed to support the change between Legacy and APF format
inside a pcap or during a stream.

This is how it works :
The switch is detected in the IsLidarPacket function of the Meta interpreter.
As soon as the first switch is detected, the Meta (after saving the previous calib file)
invokes the event "FirstSwitchOfInterpreterEvent".
This allows to this behavior to get the interpreter affected by this switch.
We are in the interpreter thread at this point so it is not safe to open a QDialog (see comments in the class).
So a second signal is sent so we can open the Calibration dialog safely in the ui thread.