[fix] Fix the default source shown is spreadsheet to None
Before this commit :
The RPM was shown by default in the spreadsheet.
It forced the default "Attribute" to "row data" for other futur empty source.
So when we created a stream (which didn't receiving any data), the attribute was set to "row data" in the spreadsheet.
When data arrived, the stream fill the attribute "Point Data", and let empty "raw data".
The user thought that the spreadsheet was not working because there was no data displayed.

We need to set the active source to "None" (as it is done for the measurement grid)
so the RPM is not displayed in the spreadsheet anymore.
The default attribute is then "Point Data"